Moving toolbar-hotkeys to toolbar plugin.
