+++ src/components/SlateEditor/RichTextEditor.jsx
@@ -11,29 +11,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Editor } from 'slate-react';
-import { isCodeHotkey, isKeyHotkey } from 'is-hotkey';
+import isHotkey from 'is-hotkey';
 import BEMHelper from 'react-bem-helper';
 import { css } from '@emotion/core';
 import createSlateStore, { setSubmitted } from './createSlateStore';
-import { handleClickBlock, handleClickMark, handleClickInline } from './utils/handleMenuClicks';
 import { PluginShape } from '../../shapes';
-
-const isBoldHotkey = isKeyHotkey('mod+b');
-const isCodeHotKey = isKeyHotkey('mod+k');
-const isConceptBlockHotKey = isCodeHotkey('mod+alt+c');
-const isH2HotKey = isKeyHotkey('mod+2');
-const isH3HotKey = isKeyHotkey('mod+3');
-const isFootnoteHotKey = isCodeHotkey('mod+alt+f');
-const isItalicHotKey = isKeyHotkey('mod+i');
-const isLetteredListHotKey = isCodeHotkey('mod+alt+a');
-const isLinkHotKey = isCodeHotkey('mod+alt+l');
-const isListHotKey = isKeyHotkey('mod+l');
-const isMathHotKey = isKeyHotkey('mod+m');
-const isNumberedListHotKey = isCodeHotkey('mod+alt+1');
-const isQuoteHotKey = isCodeHotkey('mod+alt+q');
-const isSaveHotKey = isKeyHotkey('mod+s');
-const isSubHotKey = isCodeHotkey('mod+alt+s');
-const isSupHotKey = isCodeHotkey('mod+alt+h');
 
 export const classes = new BEMHelper({
   name: 'editor',
@@ -83,41 +65,10 @@ const RichTextEditor = class extends React.PureComponent {
   }
 
   onKeyDown(e, editor, next) {
-    let mark;
-    let block;
-    let inline;
     const { value } = editor;
-
-    if (isBoldHotkey(e)) {
-      mark = 'bold';
-    } else if (isCodeHotKey(e)) {
-      mark = 'code';
-    } else if (isConceptBlockHotKey(e)) {
-      inline = 'concept';
-    } else if (isFootnoteHotKey(e)) {
-      inline = 'footnote';
-    } else if (isH2HotKey(e)) {
-      block = 'heading-two';
-    } else if (isH3HotKey(e)) {
-      block = 'heading-three';
-    } else if (isItalicHotKey(e)) {
-      mark = 'italic';
-    } else if (isLetteredListHotKey(e)) {
-      block = 'letter-list';
-    } else if (isLinkHotKey(e)) {
-      inline = 'link';
-    } else if (isListHotKey(e)) {
-      block = 'bulleted-list';
-    } else if (isMathHotKey(e)) {
-      inline = 'mathml';
-    } else if (isNumberedListHotKey(e)) {
-      block = 'numbered-list';
-    } else if (isQuoteHotKey(e)) {
-      block = 'quote';
-    } else if (isSubHotKey(e)) {
-      mark = 'sub';
-    } else if (isSupHotKey(e)) {
-      mark = 'sup';
+    if (isHotkey('mod+s', e)) {
+      e.preventDefault();
+      this.props.handleSubmit();
     } else if (e.key === 'Backspace') {
       const { removeSection, index } = this.props;
       if (removeSection) {
@@ -134,20 +85,8 @@ const RichTextEditor = class extends React.PureComponent {
         }
         next();
       }
-    } else if (isSaveHotKey(e)) {
-      e.preventDefault();
-      this.props.handleSubmit();
     }
-
-    if (mark) {
-      handleClickMark(e, editor, mark);
-    } else if (block) {
-      handleClickBlock(e, editor, block);
-    } else if (inline) {
-      handleClickInline(e, editor, inline);
-    } else {
-      next();
-    }
+    next();
   }
 
   render() {
